+++ Sources/SwiftUIKit/Lists/ListButtonStyle.swift
@@ -9,28 +9,46 @@
 import SwiftUI
 
 /**
- This style makes a button take up an entire row and applies
+ This style makes buttons take up the entire row and applies
  a content shape that makes the entire view tappable.
 
  You can apply the style with `.buttonStyle(.list)`, and can
  apply it to an entire list, just as with other styles.
  */
 public struct ListButtonStyle: ButtonStyle {
+    
+    /// Create a custom style.
+    ///
+    /// - Parameters:
+    ///   - pressedOpacity: The opacity to apply when the button is pressed, by default `0.5`.
+    public init(
+        pressedOpacity: Double = 0.5
+    ) {
+        self.pressedOpacity = pressedOpacity
+    }
+    
+    /// The opacity to apply when the button is pressed.
+    var pressedOpacity: Double
 
     public func makeBody(configuration: Configuration) -> some View {
         configuration.label
             .frame(maxWidth: .infinity, alignment: .leading)
             .contentShape(Rectangle())
+            .opacity(configuration.isPressed ? pressedOpacity : 1)
     }
 }
 
 public extension ButtonStyle where Self == ListButtonStyle {
 
-    /**
-     This style makes the button label take up all available
-     width makes the entire view tappable.
-     */
+    /// The standard list card button style.
     static var list: ListButtonStyle { .init() }
+    
+    /// A custom list card button style.
+    static func list(
+        pressedOpacity: Double
+    ) -> Self {
+        .init(pressedOpacity: pressedOpacity)
+    }
 }
 
 #if os(iOS) || os(macOS)
@@ -45,6 +63,7 @@ public extension ButtonStyle where Self == ListButtonStyle {
             List {
                 ForEach(0...100, id: \.self) {
                     button("Button \($0)")
+                        .buttonStyle($0 == 0 ? .list : .list(pressedOpacity: 0.1))
                 }
             }
             .overlay(overlay)