+++ Sources/SwiftUIKit/Lists/ListCard.swift
@@ -1,6 +1,6 @@
 //
 //  ListCard.swift
-//  KeyboardKit
+//  SwiftUIKit
 //
 //  Created by Daniel Saidi on 2023-04-26.
 //  Copyright © 2023-2024 Daniel Saidi. All rights reserved.
@@ -8,44 +8,32 @@
 
 import SwiftUI
 
-/**
- This view can be used as a floating card in lists and grids.
-
- The card can be styled with a corner radius and shadow, and
- can also be provided with a list of context menu items.
- */
+/// This view can be used as floating cards in lists & grids.
+///
+/// The card can be styled with a corner radius and a shadow,
+/// and can be provided with a list of context menu items.
 public struct ListCard<Content: View, ContextMenuView: View>: View {
 
-    /**
-     Create a list card.
-
-     - Parameters:
-       - style: The style to apply, by default ``ListCardStyle/standard``.
-       - content: The card content.
-     */
+    /// Create a list card.
+    ///
+    /// - Parameters:
+    ///   - content: The card content.
     public init(
-        style: ListCardStyle = .standard,
         @ViewBuilder content: @escaping ContentBuilder
     ) where ContextMenuView == EmptyView {
-        self.style = style
         self.content = content
         self.contextMenu = { EmptyView() }
     }
 
-    /**
-     Create a list card with a context menu.
-
-     - Parameters:
-       - style: The style to apply, by default ``ListCardStyle/standard``.
-       - content: The card content.
-       - contextMenu: The card context menu content.
-     */
+    /// Create a list card with a context menu.
+    ///
+    /// - Parameters:
+    ///   - content: The card content.
+    ///   - contextMenu: The card context menu content.
     public init(
-        style: ListCardStyle = .standard,
         @ViewBuilder content: @escaping ContentBuilder,
         @ViewBuilder contextMenu: @escaping ContextMenuBuilder
     ) {
-        self.style = style
         self.content = content
         self.contextMenu = contextMenu
     }
@@ -53,13 +41,14 @@ public struct ListCard<Content: View, ContextMenuView: View>: View {
     public typealias ContentBuilder = () -> Content
     public typealias ContextMenuBuilder = () -> ContextMenuView
 
-    private let style: ListCardStyle
-
     @ViewBuilder
     private let content: ContentBuilder
 
     @ViewBuilder
     private let contextMenu: ContextMenuBuilder
+    
+    @Environment(\.listCardStyle)
+    private var style
 
     public var body: some View {
         content()
@@ -75,85 +64,6 @@ public struct ListCard<Content: View, ContextMenuView: View>: View {
     }
 }
 
-/**
- This style can be used to style a ``ListCard``.
- */
-public struct ListCardStyle {
-
-    /**
-     Create a list card style
-
-     - Parameters:
-       - cornerRadius: The corner radius to apply, by default `8.0`.
-       - shadowStyle: The shadowStyle to apply, by default ``ViewShadowStyle/listCard``.
-     */
-    public init(
-        cornerRadius: Double = 8.0,
-        shadowStyle: ViewShadowStyle = .listCard
-    ) {
-        self.cornerRadius = cornerRadius
-        self.shadowStyle = shadowStyle
-    }
-
-    /// The corner radius to apply.
-    public var cornerRadius: Double
-
-    /// The shadow style to apply.
-    public var shadowStyle: ViewShadowStyle
-}
-
-public extension ListCardStyle {
-
-    /**
-     The standard list card style.
-
-     You can change this style to affect the standard global
-     style of the ``ListCard`` view.
-     */
-    static var standard = ListCardStyle()
-}
-
-/**
- This button styles can be used to scale down a ``ListCard``.
- */
-public struct ListCardButtonStyle: ButtonStyle {
-
-    /**
-     Create a list card button style
-
-     - Parameters:
-       - pressedScale: The scale to apply when the button is pressed, by default `0.98`.
-     */
-    public init(pressedScale: Double = 0.98) {
-        self.pressedScale = pressedScale
-    }
-
-    /// The scale to apply when the button is pressed.
-    public var pressedScale: Double
-
-    public func makeBody(configuration: Configuration) -> some View {
-        configuration.label
-            .scaleEffect(configuration.isPressed ? pressedScale : 1)
-    }
-}
-
-public extension ButtonStyle where Self == ListCardButtonStyle {
-
-    /**
-     The standard list card button style.
-
-     You can change this style to affect the standard global
-     style of the ``ListCard`` button view.
-     */
-    static var listCard: ListCardButtonStyle { .init() }
-
-    /**
-     A list card button style.
-     */
-    static func listCard(pressedScale: Double) -> ListCardButtonStyle {
-        .init(pressedScale: pressedScale)
-    }
-}
 
 public extension ViewShadowStyle {
 
@@ -170,16 +80,32 @@ public extension ViewShadowStyle {
 
 #Preview {
     
-    Button {
-    } label: {
-        ListCard {
-            Color.red.frame(width: 200, height: 200)
-        } contextMenu: {
-            
+    VStack {
+        Group {
+            Button {} label: {
+                ListCard {
+                    Color.red.frame(width: 200, height: 200)
+                }
+            }
+            .buttonStyle(
+                .listCard(
+                    animation: .bouncy,
+                    pressedScale: 0.2
+                )
+            )
+            Button {} label: {
+                ListCard {
+                    Color.red.frame(width: 200, height: 200)
+                } contextMenu: {
+                    Button {} label: {
+                        Text("Press me")
+                    }
+                }
+            }
         }
+        .buttonStyle(.listCard)
+        .padding(50)
+        .background(Color.gray)
+        .cornerRadius(20)
     }
-    .buttonStyle(.listCard)
-    .padding(50)
-    .background(Color.gray)
-    .cornerRadius(20)
 }